Move node input into MeshReader
